cleanup - explicit rename sharedkernel package into common
