+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/domain/common/Currency.kt
@@ -1,0 +1,7 @@
+package github.com.hukuta94.delivery.core.domain.common
+
+enum class Currency {
+    RUB,
+    KZT,
+    BONUS,
+}
