+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/domain/common/Distance.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.core.domain.sharedkernel
+package github.com.hukuta94.delivery.core.domain.common
 
 import kotlin.math.abs
