+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/domain/order/Order.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.core.domain.order
 
 import github.com.hukuta94.delivery.core.domain.Aggregate
 import github.com.hukuta94.delivery.core.domain.courier.Courier
-import github.com.hukuta94.delivery.core.domain.sharedkernel.Location
+import github.com.hukuta94.delivery.core.domain.common.Location
 import java.util.UUID
 
 class Order internal constructor(
