--- core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/domain/sharedkernel/Currency.kt
@@ -1,7 +1,0 @@
-package github.com.hukuta94.delivery.core.domain.sharedkernel
-
-enum class Currency {
-    RUB,
-    KZT,
-    BONUS,
-}
