+++ core/test-fixtures/src/main/kotlin/github/com/hukuta94/delivery/core/domain/common/LocationFixtures.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.core.domain.sharedkernel
+package github.com.hukuta94.delivery.core.domain.common
 
 fun newLocation(abscissa: Int, ordinate: Int) = Location(abscissa, ordinate)
