+++ core/test-fixtures/src/main/kotlin/github/com/hukuta94/delivery/core/domain/order/OrderFixtures.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.core.domain.order
 
 import github.com.hukuta94.delivery.core.domain.courier.Courier
 import github.com.hukuta94.delivery.core.domain.courier.newCourier
-import github.com.hukuta94.delivery.core.domain.sharedkernel.Location
+import github.com.hukuta94.delivery.core.domain.common.Location
 import java.util.*
 
 fun newOrder(
