+++ core/tests/src/test/kotlin/github/com/hukuta94/delivery/core/domain/common/MoneyTest.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.core.domain.sharedkernel
+package github.com.hukuta94.delivery.core.domain.common
 
 import github.com.hukuta94.delivery.core.port.Exchange
 import io.kotlintest.shouldBe
