+++ core/tests/src/test/kotlin/github/com/hukuta94/delivery/core/domain/order/OrderTest.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.core.domain.order
 
 import github.com.hukuta94.delivery.core.domain.courier.newCourier
-import github.com.hukuta94.delivery.core.domain.sharedkernel.Location
+import github.com.hukuta94.delivery.core.domain.common.Location
 import io.kotlintest.assertSoftly
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldNotBe
