+++ infrastructure/src/main/kotlin/github/com/hukuta94/delivery/infrastructure/adapter/orm/model/entity/CourierJpaEntity.kt
@@ -4,7 +4,7 @@ import github.com.hukuta94.delivery.core.domain.courier.Courier
 import github.com.hukuta94.delivery.core.domain.courier.CourierName
 import github.com.hukuta94.delivery.core.domain.courier.CourierStatus
 import github.com.hukuta94.delivery.core.domain.courier.Transport
-import github.com.hukuta94.delivery.core.domain.sharedkernel.Location
+import github.com.hukuta94.delivery.core.domain.common.Location
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.model.converter.CourierNameConverter
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.model.converter.CourierStatusConverter
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.model.converter.LocationConverter
